+++ scrapy_climate/spider.py
@@ -1,119 +1,5 @@
 # -*- coding: utf-8 -*-
-
-import scrapy
-
-from urllib.parse import urlparse, urlunparse
-from scrapy.http import Response, Request
-from scrapy.selector import SelectorList
-
-from .items import ArticleItem
-from .tools import convert_list_to_string, fetch_scraped_indexes
-
-
-class ParserMixin:
-
-    # CSS selectors for finding "news-list tag"/"article tag"
-    # on "news-list page"/"article page" in the response
-    # Must be a string.
-    _css_selector_news_list = None
-    _css_selector_article = None
-
-    # XPATH selector used to find path or url to "article page" under
-    # "news-list tag". It is relative selector.
-    # Must be a string.
-    _xpath_selector_path_or_url = None
-
-    # XPATH selector lists used to find needed data on the "article page"
-    # under "article tag". There are relative selectors.
-    # Must be a string.
-    _xpath_selector_list_tags = None
-    _xpath_selector_list_text = None
-    _xpath_selector_list_header = None
-
-    # ================
-    #  "find" methods
-    # ================
-    # uses `SelectorList` to find specific HTML tag on the page or inside
-    # another HTML tag
-    def _find_article_in_response(self, response: Response) -> SelectorList:
-        return response.css(self._css_selector_article)
-
-    def _find_news_list_in_response(self, response: Response) -> SelectorList:
-        return response.css(self._css_selector_news_list)
-
-    def _find_tags_in_article(self, article: SelectorList) -> SelectorList:
-        return self._find_by_xpath_list(article,
-                                        self._xpath_selector_list_tags)
-
-    def _find_text_in_article(self, article: SelectorList) -> SelectorList:
-        return self._find_by_xpath_list(article,
-                                        self._xpath_selector_list_text)
-
-    def _find_header_in_article(self, article: SelectorList) -> SelectorList:
-        return self._find_by_xpath_list(article,
-                                        self._xpath_selector_list_header)
-
-    # ===================
-    #  "extract" methods
-    # ===================
-    # uses `SelectorList.extract` method to extract data from HTML tag
-    def _extract_tags(self, article: SelectorList) -> str:
-        return convert_list_to_string(
-            self._find_tags_in_article(article).extract(), ',')
-
-    def _extract_text(self, article: SelectorList) -> str:
-        return convert_list_to_string(
-            self._find_text_in_article(article).extract(),
-            separator='',
-            handler=self._clear_text_field,)
-
-    def _extract_header(self, article: SelectorList) -> str:
-        return self._find_header_in_article(article).extract_first()
-
-    def _extract_url_or_path_from_block(self, block: SelectorList) -> str:
-        return block.xpath(self._xpath_selector_path_or_url).extract_first()
-
-    # =========
-    #  helpers
-    # =========
-    @staticmethod
-    def _find_by_xpath_list(
-            article: SelectorList,
-            xpath_string_selectors_list: list or tuple) -> SelectorList:
-        """
-        Iterates over XPATH string selectors in
-        `xpath_string_selector_list` uses them for `article` selector and
-        returns SelectorList of results.
-        :param article: SelectorList that matches "article tag"
-        :param xpath_string_selectors_list: list or tuple of string XPATH
-        selectors for `article` selector
-        :return: SelectorList with selector fo every given string selector
-        in `xpath_string_selector_list` in `article` selector
-        """
-        return SelectorList([article.xpath(selector)
-                             for selector in xpath_string_selectors_list])[0]
-
-    def _clear_text_field(self, text: str) -> str:
-        """
-        Function clears text from different special characters that must be
-        in the worksheet.
-        :param text: string that must be cleared from unneeded characters
-        :return: cleared string
-        """
-        return str(text).replace('\xa0', ' ').replace('\n', '')
-
-
-class TemplateSpider(scrapy.Spider, ParserMixin):
-    """
-    This class must not be used properly, only for inheritance.
-
-    It implements Scrapy spider callbacks for scraping articles from news-like
-    web-site, without duplicates (see `fetch_scraped_indexes` function in the
-    `tools` module.)
-
-    What must be implemented for usage?
-    * class fields: `name`, `_start_path`, `_start_domain`, `_scheme`
-    * methods: `_convert_path_to_index`
+""" Module for spider class templates.
 
     Entities:
     * news-list page - page on the web-site that have HTML tag (e. g.
@@ -127,6 +13,32 @@ class TemplateSpider(scrapy.Spider, ParserMixin):
     * index - part of the article page URL that can be used to identify the
     article page to not scrape it twice
     * callback - method which takes request and yields another request or item
+"""
+
+import scrapy
+
+from urllib.parse import urlparse, urlunparse
+from scrapy.http import Response, Request
+
+from .items import ArticleItem
+from .tools import fetch_scraped_indexes
+from .mixins import ParserMixin
+
+
+class TemplateSpider(scrapy.Spider, ParserMixin):
+    """
+    This class must not be used properly, only for inheritance.
+
+    It implements Scrapy spider callbacks for scraping articles from news-like
+    web-site, without duplicates (see `fetch_scraped_indexes` function in the
+    `tools` module.)
+
+    What must be implemented for usage?
+    * class fields: `name`, `_start_path`, `_start_domain`, `_scheme`
+    * class fields from mixins: `_css_selector_article`,
+    `_css_selector_news_list`, `xpath_selector_list_header`,
+    `_xpath_selector_list_tags`, `_xpath_selector_path_or_url`
+    * methods: `_convert_path_to_index`
 
     How it works at all:
     1. first request is to `start_urls`'s first (and only) URL, that depends
